ForwardBot:
Initial commit of ForwardBot

A bot intended to allow for setup of forwarding rules for channels, e.g.
to save attachments in an art channel or move images to another channel
when uploaded.

It's purpose at first was primarily to allow for archiving of fan art
and other files that a server I was on was doing.

Initial commit consists of:

Basic forwarding setup functionality and rules
Packaging of Python package with dependencies
ReadMe with example uses